+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The most recent changes are on top, in each type of changes category.
 
 ### Added
 
+- Scene metadata search [#422](https://github.com/cyfronet-fid/sat4envi/issues/422)
 - Listen to scene files notifications from AMQP queue [#440](https://github.com/cyfronet-fid/sat4envi/issues/440)
 - Add version display for provided env variable during build time [#459](https://github.com/cyfronet-fid/sat4envi/issues/459)
 - Include commit SHA in maven project.version and pass it to npm [#461](https://github.com/cyfronet-fid/sat4envi/issues/461)